+++ common/src/main/java/dev/zoty/maxinuta/platform/MaxinutaPlatformHelper.java
@@ -1,5 +1,4 @@
 package dev.zoty.maxinuta.platform;
 
 public interface MaxinutaPlatformHelper {
-
 }