+++ forge/src/main/java/dev/zoty/maxinuta/forge/MaxinutaForge.java
@@ -1,5 +1,6 @@
-package dev.zoty.maxinuta;
+package dev.zoty.maxinuta.forge;
 
+import dev.zoty.maxinuta.Maxinuta;
 import dev.zoty.maxinuta.item.MaxinutaCreativeTabs;
 import dev.zoty.maxinuta.item.MaxinutaItems;
 import dev.zoty.maxinuta.registry.internal.RegistrationFunction;
@@ -7,22 +8,13 @@ import net.minecraft.core.registries.Registries;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.registries.RegisterEvent;
 
 import java.util.function.Consumer;
 
 @Mod(Maxinuta.MOD_ID)
 public class MaxinutaForge {
-    
     public MaxinutaForge() {
-        IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
-    
-        // This method is invoked by the Forge mod loader when it is ready
-        // to load your mod. You can access Forge and Common code in this
-        // project.
-    
-        // Use Forge to bootstrap the Common mod.
         Maxinuta.LOG.info("Hello Forge world!");
         Maxinuta.init();
     }