+++ quilt/src/main/java/dev/zoty/maxinuta/quilt/platofrm/QuiltPlatformHelper.java
@@ -1,0 +1,21 @@
+package dev.zoty.maxinuta.quilt.platofrm;
+
+import dev.zoty.maxinuta.platform.services.IPlatformHelper;
+import org.quiltmc.loader.api.QuiltLoader;
+
+public class QuiltPlatformHelper implements IPlatformHelper {
+    @Override
+    public String getPlatformName() {
+        return "Quilt";
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+        return QuiltLoader.isModLoaded(modId);
+    }
+
+    @Override
+    public boolean isDevelopmentEnvironment() {
+        return QuiltLoader.isDevelopmentEnvironment();
+    }
+}